Jumppack base
Changed the way healing is dealt - it is now ignoring the "min 25% of damage dealt" rule. All current healing is dealt as Toxic damage (with minus damage), so toxic resistances affect the healing.
Fixed healing, it was applied 2 times for server player
Started working on a jumppack - added it as an equippable item. Had to make a workaround for customization system. Jumppack is an equipment AND customization item, so had to add a cross-support. Will have to lock backpacks for customization if such items are equipped.
Made root motion animation for jump pack.
Added effects and changed evade logic for marines with jump pack.